Add skiko WinUI backend foundation
Adds the skiko-winui module, WinUI JVM rendering/input/accessibility foundations, samples, validation script, and PLAN tracking. Records current winui-mingw blocker in kotlin-winrt Native compiler-plugin lowering while JVM core and smoke validation pass with windows.foundation geometry projections.

+++ skiko/src/nativeMain/kotlin/org/jetbrains/skiko/Resources.native.kt
@@ -13,7 +13,7 @@ actual suspend fun loadBytesFromPath(path: String): ByteArray {
 
     val size = file.let {
         fseek(it, 0, SEEK_END)
-        val size = ftell(it)
+        val size = ftell(it).toLong()
         fseek(it, 0, SEEK_SET)
         size
     }
